% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayCurvesRecur.R
\name{DelayCurvesRecur}
\alias{DelayCurvesRecur}
\title{DelayCurvesRecur}
\usage{
DelayCurvesRecur(basic, advanced)
}
\arguments{
\item{basic}{A list containing basic input parameters validated by the DelayInputValidator function}

\item{advanced}{A list containing advanced input parameters (e.g., computational settings) validated by the DelayInputValidator function}
}
\value{
A list containing estimated model details
}
\description{
Implements the Stage II recursion
}
