% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayFigure.R
\name{DelayFigure}
\alias{DelayFigure}
\title{DelayFigure}
\usage{
DelayFigure(
  basic,
  mat,
  xlims = NULL,
  xnbreaks = NULL,
  ylims = NULL,
  ynbreaks = NULL,
  thresholds = TRUE,
  points = TRUE
)
}
\arguments{
\item{basic}{A list containing basic input parameters validated by the DelayInputValidator function}

\item{mat}{A list containing estimated model details created by the DelayStageOne function}

\item{xlims}{Limits for the horizontal axis}

\item{xnbreaks}{Number of breaks for the horizontal axis}

\item{ylims}{Limits for the vertical axis}

\item{ynbreaks}{Number of breaks for the vertical axis}

\item{thresholds}{If TRUE, plot threshold boundaries for a one stage trial}

\item{points}{If TRUE, plot points for a series of one stage trials}
}
\value{
A list containing estimated model details
}
\description{
Visualises stopping boundaries
}
