% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayInputConstructor.R
\name{DelayInputConstructor}
\alias{DelayInputConstructor}
\title{DelayInputConstructor}
\usage{
DelayInputConstructor(basicarray = list(), advancedarray = list())
}
\arguments{
\item{basicarray}{A list containing any basic input parameters}

\item{advancedarray}{A list containing any advanced input parameters (e.g., computational settings)}
}
\value{
A list containing the initial basic and advanced data structures
}
\description{
Creates the initial basic and advanced data structures which are used as inputs for the model
}
