% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayInputValidator.R
\name{DelayInputValidator}
\alias{DelayInputValidator}
\title{DelayInputValidator}
\usage{
DelayInputValidator(basic, advanced)
}
\arguments{
\item{basic}{A list containing basic input parameters created by the DelayInputConstructor function}

\item{advanced}{A list containing advanced input parameters (e.g., computational settings) created by the DelayInputConstructor function}
}
\value{
A list containing validated basic and advanced data structures and an error list containing messages if any errors were found
}
\description{
Checks the basic and advanced data structures for validity
}
