% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayStageOne.R
\name{DelayStageOne}
\alias{DelayStageOne}
\title{DelayStageOne}
\usage{
DelayStageOne(basic, advanced, mat)
}
\arguments{
\item{basic}{A list containing basic input parameters validated by the DelayInputValidator function}

\item{advanced}{A list containing advanced input parameters (e.g., computational settings) cvalidated by the DelayInputValidator function}

\item{mat}{A list containing estimated model details created by the DelayCurvesRecur function}
}
\value{
A list containing a modified copy of the mat input list
}
\description{
Implements the Stage I recursion
}
