% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htadelayR-package.R
\docType{package}
\name{htadelayR-package}
\alias{htadelayR}
\alias{htadelayR-package}
\title{htadelayR: Sequential Experimentation with Delayed Response - An R Implementation}
\description{
The code computes optimal stopping times for sequential experiments when samples are observed with delay.
}
\author{
\strong{Maintainer}: Michael Dymock \email{michael.dymock1@hotmail.com} (\href{https://orcid.org/0000-0003-4881-2860}{ORCID})

Authors:
\itemize{
  \item Stephen E. Chick
}

}
\keyword{internal}
